backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf2
sentence-transformers
faiss-cpu
groq
python-dotenv
pydantic
numpy
